LayoutTests/http/tests/media/media-source/video-media-source-async-events-expected.txt

Verify that all MediaSource & SourceBufferList events are dispatched asynchronously.

onSourceOpen
onAddSourceBuffer
onAddSourceBuffer
onSourceEnded
onRemoveSourceBuffer
onRemoveSourceBuffer
onSourceClose
END OF TEST
